+++ R/graph-functions.R
@@ -482,9 +482,12 @@ insert_one_edge <- function (graph, index, x, y, gr_cols) {
         y = graph [[gr_cols$yto]] [index + 1]
     )
     if (is_graph_spatial (graph)) {
-        requireNamespace ("geodist")
-        d1 <- geodist::geodist (xy1, c (x = x, y = y), measure = "geodesic")
-        d2 <- geodist::geodist (xy2, c (x = x, y = y), measure = "geodesic")
+        # Get geodist measure, noting that graph has no hash at this stage, so
+        # full calculation will be executed.
+        measure <- get_geodist_measure (graph)
+
+        d1 <- geodist::geodist (xy1, c (x = x, y = y), measure = measure)
+        d2 <- geodist::geodist (xy2, c (x = x, y = y), measure = measure)
     } else {
         d1 <- sqrt ((xy1 [1] - x)^2 + (xy1 [2] - y)^2)
         d2 <- sqrt ((x - xy2 [1])^2 + (y - xy2 [2])^2)